fix: race condition with external binaries and files with the same content
https://github.com/phoenixframework/phoenix/pull/5246 causes Phoenix to
run asset compression in parallel. If some of the files have the same
contents, creating a filename based on the md5 of the content is not
unique anymore. This results in multiple processes working on the
same file at once.

The issue is easily noticable, since `File.rm` complains if the
file doesn't exist. This is not an artifical example, since auto
generated OpenStreetMap vector tiles are often empty and thus
also have the same contents.